refactor of directory structure
