+++ .gitignore
@@ -4,3 +4,4 @@ build.sh
 path.go
 rssagg
 scrapperscratch.md
+.idea
