+++ go.mod
@@ -1,13 +1,10 @@
-module rssagg
+module github.com/etrinque/rssagg
 
-go 1.22.0
+go 1.23.0
 
 require (
 	github.com/go-chi/chi v1.5.5
-	github.com/joho/godotenv v1.5.1
-)
-
-require (
 	github.com/google/uuid v1.6.0
+	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 )
