+++ utils/handlers.go
@@ -3,8 +3,8 @@ package utils
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/etrinque/rssagg/database"
 	"net/http"
-	"rssagg/internal/database"
 	"strconv"
 	"time"
