+++ utils/scrapper.go
@@ -5,10 +5,10 @@ import (
 	"database/sql"
 	"encoding/xml"
 	"errors"
+	"github.com/etrinque/rssagg/database"
 	"io"
 	"log"
 	"net/http"
-	"rssagg/internal/database"
 	"strings"
 	"sync"
 	"time"
